+++ src/components/Bills/useBillForm.ts
@@ -7,13 +7,11 @@ import { Vendor } from '../../types/vendors'
 export type BillForm = {
   bill_number?: string
   vendor?: Vendor
-  vendor_address?: string // read-only
-  received_at?: string // OR DATE
-  due_date?: string // OR DATE
+  vendor_address?: string
+  received_at?: string
+  due_date?: string
   terms?: string
   line_items?: Partial<BillLineItem>[]
-  tax?: boolean // not supported by API?
-  billable?: boolean // not supported by API?
 }
 
 export const useBillForm = (bill: Bill) => {
@@ -26,10 +24,8 @@ export const useBillForm = (bill: Bill) => {
       vendor_address: bill.vendor?.address_string,
       received_at: bill.received_at,
       due_date: bill.due_at,
-      terms: '', // @TODO match to updated API
+      terms: '',
       line_items: bill.line_items,
-      // tax: false, // @TODO match to updated API
-      // billable: false, // @TODO match to updated API
     },
     onSubmit: async ({ value }) => {
       console.log('onSubmit - sending...', value)
