+++ src/components/Button/SubmitButton.tsx
@@ -14,8 +14,8 @@ export interface SubmitButtonProps
   disabled?: boolean
   error?: boolean | string
   active?: boolean
-  variant?: ButtonProps['variant']
   iconOnly?: boolean
+  variant?: ButtonProps['variant']
   action?: SubmitAction
   noIcon?: boolean
   tooltip?: ButtonProps['tooltip']
