+++ src/styles/inputs.scss
@@ -123,7 +123,6 @@
 
 .Layer__amount-input {
   text-align: right;
-  font-size: 20px;
 }
 
 .Layer__textarea {
